megadetector/requirements.txt
matplotlib==3.10.1
numpy==2.2.4
Pillow==11.1.0
PytorchWildlife==1.2.0
gradio
